Add files via upload
This was an analysis I worked on that looked at the migration patterns of retirement age (65+) New Yorkers and where in the country they move to. Along with the analysis, I conducted research aimed to convince retirement age individuals that the ideal place to retire to in America is New York City.